backend/requirements.txt
# Core Web Framework
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0  # Use SQLite for simplicity

# Validation
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# AI & Vector Database
qdrant-client==1.12.1
google-generativeai==0.8.4  # Google's official SDK
openai==1.54.5  # OpenAI SDK configured for Gemini compatibility

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dotenv==1.0.1
numpy==2.1.3
pillow==11.0.0
tiktoken==0.8.0
